+++ crates/cli/src/commands/node/run.rs
@@ -78,7 +78,7 @@ pub struct RunOpts {
 
     /// Enables the UI for the node
     #[clap(long, action, default_value = "false")]
-    pub enable_ui: bool,
+    pub gui: bool,
 
     /// Disables networking capabilities of the node
     #[clap(long, action, default_value = "false")]
@@ -136,7 +136,7 @@ impl From<RunOpts> for NodeConfig {
             // a hack, but it works for now.
             keypair: default_node_config.keypair,
             disable_networking: opts.disable_networking,
-            enable_ui: opts.enable_ui,
+            gui: opts.gui,
             rendezvous_server_address: opts.rendezvous_server_address,
             public_ip_address: opts.raptorq_gossip_address,
         }
@@ -165,7 +165,7 @@ impl Default for RunOpts {
             bootstrap_node_addresses: Default::default(),
             http_api_title: Default::default(),
             http_api_version: Default::default(),
-            enable_ui: Default::default(),
+            gui: Default::default(),
             disable_networking: Default::default(),
             rendezvous_local_address: ipv4_localhost_with_random_port,
             rendezvous_server_address: ipv4_localhost_with_random_port,
@@ -253,7 +253,7 @@ impl RunOpts {
             http_api_address: other.http_api_address,
             http_api_title,
             http_api_version,
-            enable_ui: false,
+            gui: false,
             disable_networking: false,
             rendezvous_local_address: other.rendezvous_local_address,
             rendezvous_server_address: other.rendezvous_server_address,
