+++ crates/node/src/node.rs
@@ -52,6 +52,7 @@ pub struct Node {
     raptor_handle: RaptorHandle,
     scheduler_handle: SchedulerHandle,
     grpc_server_handle: RuntimeHandle,
+    node_gui_handle: RuntimeHandle,
 }
 
 pub type UnboundedControlEventReceiver = UnboundedReceiver<Event>;
@@ -64,16 +65,6 @@ impl Node {
     ) -> Result<Self> {
         // Copy the original config to avoid overwriting the original
         let mut config = config.clone();
-
-        if config.enable_ui {
-            match Node::configure_node_ui(config.clone()) {
-                Ok(_) => info!("Node UI is ready"),
-                Err(e) => {
-                    info!("Failed to start Node UI: {}", e);
-                    info!("Node UI will not be available");
-                }
-            }
-        }
 
         let vm = None;
         let keypair = config.keypair.clone();
@@ -115,6 +106,7 @@ impl Node {
             raptor_handle: runtime_components.raptor_handle,
             scheduler_handle: runtime_components.scheduler_handle,
             grpc_server_handle: runtime_components.grpc_server_handle,
+            node_gui_handle: runtime_components.node_gui_handle,
         })
     }
 
@@ -182,6 +174,11 @@ impl Node {
             info!("rpc server shut down");
         }
 
+        if let Some(handle) = self.node_gui_handle {
+            handle.await??;
+            info!("node gui shut down");
+        }
+
         self.router_handle.await?;
 
         info!("node shutdown complete");
@@ -213,67 +210,6 @@ impl Node {
     /// Returns the node's type
     pub fn node_type(&self) -> NodeType {
         self.config.node_type
-    }
-
-    fn configure_node_ui(config: NodeConfig) -> anyhow::Result<()> {
-            info!("Configuring Node {}", &config.id);
-            info!("Ensuring environment has required dependencies");
-
-            match Command::new("npm").args(&["version"]).status() {
-                Ok(_) => info!("NodeJS is installed"),
-                Err(e) => {
-                    return Err(NodeError::Other(format!("NodeJS is not installed: {}", e)).into());
-                }
-            }
-
-            info!("Ensuring yarn is installed");
-
-            match Command::new("yarn").args(&["--version"]).status() {
-                Ok(_) => info!("Yarn is installed"),
-                Err(e) => {
-                    let install_yarn = Command::new("npm")
-                        .args(&["install", "-g", "yarn"])
-                        .current_dir("infra/ui")
-                        .output();
-
-                    match install_yarn {
-                        Ok(_) => (),
-                        Err(_) => {
-                            return Err(NodeError::Other(format!("Failed to install yarn: {}", e)).into());
-                        }
-                    }
-                }
-            }
-
-            info!("Installing dependencies");
-
-            match Command::new("yarn")
-                .args(&["install"])
-                .current_dir("infra/ui")
-                .status()
-            {
-                Ok(_) => info!("Dependencies installed successfully"),
-                Err(e) => {
-                    return Err(NodeError::Other(format!("Failed to install dependencies: {}", e)).into());
-                }
-            }
-
-            info!("Spawning UI");
-
-            match Command::new("yarn")
-                .args(&["dev"])
-                .current_dir("infra/ui")
-                .spawn()
-            {
-                Ok(_) => info!("UI spawned successfully"),
-                Err(e) => {
-                    return Err(NodeError::Other(format!("Failed to spawn UI: {}", e)).into());
-                }
-            }
-
-            info!("Finished spawning UI");
-
-        Ok(())
     }
 
     #[deprecated(note = "use node_type instead")]
